Refactor LocalManager into FileInspector for package type checks and XdgIntegration for xdg asset management
